+++ chronosphere/tfschema/log_search_filter.go
@@ -2,7 +2,7 @@ package tfschema
 
 import "github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 
-var LogSearchFilterSchema = map[string]*schema.Schema{
+var LogSearchSchema = map[string]*schema.Schema{
 	"query": {
 		Type:     schema.TypeString,
 		Required: true,