+++ examples/log-allocation-config/main.tf
@@ -4,10 +4,10 @@ resource "chronosphere_log_allocation_config" "my_log_allocation_config" {
       percent_of_license = 90.1
     }
     priorities {
-      high_priority_filters {
+      high_priority_filter {
         query = "severity='error'"
       }
-      low_priority_filters {
+      low_priority_filter {
         query = "severity='debug'"
       }
     }
@@ -19,10 +19,10 @@ resource "chronosphere_log_allocation_config" "my_log_allocation_config" {
       percent_of_license = 9.9
     }
     priorities {
-      high_priority_filters {
+      high_priority_filter {
         query = "severity='error'"
       }
-      low_priority_filters {
+      low_priority_filter {
         query = "severity='debug'"
       }
     }